fix bugs
## Changes Made
- change serving size to allow fractions
- change AI validation to be lighter and not too strict when adding ingredients

+++ lib/features/explore/domain/entities/explore_recipe.dart
@@ -22,7 +22,7 @@ class ExploreRecipe {
   final String allergenInfo;
   final String totalTime;
   final String difficulty;
-  final int servings;
+  final double servings;
   final double rating;
   final int ratingCount;
   final int commentCount;
